# apis/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "description": "Get all user order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Get all user order",
                "parameters": [
                    {
                        "description": "GetOrderRequest",
                        "name": "GetOrderRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.GetOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.GetOrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a a new customer order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Create a order",
                "parameters": [
                    {
                        "description": "CreateOrderRequest",
                        "name": "CreateOrderRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/services.CreateOrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}": {
            "put": {
                "description": "Update user order status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Update user order status",
                "parameters": [
                    {
                        "description": "UpdateStatusRequest",
                        "name": "UpdateStatusRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/services.UpdateStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}/cancel": {
            "put": {
                "description": "Cancel user order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Cancel user order",
                "parameters": [
                    {
                        "description": "CancelOrderRequest",
                        "name": "CancelOrderRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CancelOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/services.CancelOrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "enums.OrderStatus": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5
            ],
            "x-enum-varnames": [
                "Pending",
                "Confirmed",
                "Shipped",
                "Delivered",
                "Cancelled"
            ]
        },
...
        "services.CancelOrderRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "services.CancelOrderResponse": {
            "type": "object",
            "properties": {
                "orderId": {
                    "type": "string"
                }
            }
        },
        "services.CreateOrderRequest": {
            "type": "object",
            "required": [
                "orderItems",
                "userId"
            ],
            "properties": {
                "orderItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.OrderItemModel"
                    }
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "services.CreateOrderResponse": {
            "type": "object",
            "properties": {
                "orderId": {
                    "type": "string"
                }
            }
        },
...
        "services.GetOrderItemModel": {
            "type": "object",
            "properties": {
                "categoryName": {
                    "type": "string"
                },
                "orderItemId": {
                    "type": "string"
                },
                "productDescription": {
                    "type": "string"
                },
                "productName": {
                    "type": "string"
                },
                "productPrice": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "subTotal": {
                    "type": "number"
                }
            }
        },
        "services.GetOrderRequest": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "order": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "pageLength": {
                    "type": "integer"
                },
                "sortBy": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "services.GetOrderResponse": {
            "type": "object",
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.OrderModel"
                    }
                }
            }
        },
...
        "services.OrderItemModel": {
            "type": "object",
            "required": [
                "productId",
                "quantity"
            ],
            "properties": {
                "productId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "services.OrderModel": {
            "type": "object",
            "properties": {
                "orderId": {
                    "type": "string"
                },
                "orderitems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.GetOrderItemModel"
                    }
                },
                "taxAmount": {
                    "type": "number"
                },
                "totalPrice": {
                    "type": "number"
                }
            }
        },
...
        "services.UpdateStatusRequest": {
            "type": "object",
            "required": [
                "id",
                "status"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/enums.OrderStatus"
                }
            }
        },
        "services.UpdateStatusResponse": {
            "type": "object",
            "properties": {
                "orderId": {
                    "type": "string"
                }
            }
        }
    }
}